requirements.txt
altgraph==0.17.2
annotated-types==0.7.0
distlib==0.3.9
filelock==3.16.1
future==0.18.2
macholib==1.15.2
platformdirs==4.3.6
pydantic==2.10.4
pydantic_core==2.27.2
six==1.15.0
typing_extensions==4.12.2
virtualenv==20.28.0
google-generativeai
pandas
